requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
websockets>=11.0.0
python-dateutil>=2.8.2
matplotlib>=3.7.0
pandas>=2.0.0
